cleanup: DWORD usage in the application messenger, network, and resolutions (guisettings as well but that went in with r23085)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@23088 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/ApplicationMessenger.cpp
@@ -495,7 +495,7 @@ case TMSG_POWERDOWN:
 
     case TMSG_NETWORKMESSAGE:
       {
-        g_application.getNetwork().NetworkMessage((CNetwork::EMESSAGE)pMsg->dwParam1, pMsg->dwParam2);
+        g_application.getNetwork().NetworkMessage((CNetwork::EMESSAGE)pMsg->dwParam1, (int)pMsg->dwParam2);
       }
       break;
